requirements.txt
autopep8==1.7.0
celery==5.2.7
django-celery-beat==2.3.0
Django==4
isodate==0.6.1
Pillow==9.2.0
psycopg2==2.9.3
djangorestframework==3.13.1
django-filter==22.1
markdown==3.4.1
redis==4.3.4
requests==2.28.1
django-celery-results==2.4.0